requirements.txt
sqlalchemy>=2.0
httpx>=0.27
pydantic>=2.0
streamlit>=1.38
pytest>=8.0
python-dotenv>=1.0